requirements.txt
selenium==4.16.0
pytest==8.0.0
allure-pytest==2.13.2
pytest-xdist==3.5.0